[update]関数の移動 #36
